# sixth-week/swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/test/{id}": {
            "get": {
                "description": "获取用户所有的操作日志",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "操作审计"
                ],
                "summary": "获取日志列表",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "description": "页码",
                        "name": "page",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "每页多少条信息",
                        "name": "limit",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "模块 SrcType",
                        "name": "mod",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "操作 Type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.ListResponse"
                        }
                    },
                    "400": {
                        "description": "Id\":100}  \"能力模型不存在",
                        "schema": {
                            "$ref": "#/definitions/main.IDD"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.IDD": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "main.ListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.LogRecord"
                    }
                }
            }
        },
        "main.LogRecord": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "integer"
                },
                "dataName": {
                    "type": "string"
                },
                "logType": {
                    "type": "integer"
                },
                "operatorTime": {
                    "type": "integer"
                },
                "srcType": {
                    "type": "integer"
                },
                "uid": {
                    "type": "integer"
                },
                "userId": {
                    "type": "string"
                }
            }
        }
    }
}